Add Trilinear Interpolation and Hermitian Smoothing to Perlin Noise
The original C++ code performed two passes to generate the trilinear
interpolation.  First the random floats were calculated (and assigned to
a 2x2x2 matrix) and then those values were used in the accumulation.

I choose to use one pass that chooses the random floats and accumulates
in one go.

I don't know if this is faster, as the inner XOR and permutation lookup
happens on each iteration.  Benchmarking still required.