Fix Perlin to Generate Correct Marbling
Reworked Perlin noise generation to more closely match C++ reference
implementation.  Specifically, create an array of permutations BEFORE
accumulating to avoid re-indexing into self.perm_* on each loop.

I still think this is less performant, but re-examining the code did
reveal a bug in the accumulator where the dot product was backward.

Also fixed a bug in the index permutation to correctly swap over the
entire range.  Generates a true random array.

Need to re-examine in the future when looking for optimization
opportunities.