Add `BVHNode` (Bounding Volume Hierarchies) and Apply to Scene
As the scenes become more complex, a BVH is required to speed up
rendering as we want to cull rays that don't hit anything as quickly as
possible.

Using AABB bounding boxes, we can quickly check if a ray will hit
anything within a region, and if not, bail out.  If the ray does
intersect the AABB, cast into the next BVH, until an object is hit.

NOTE: Deviating from the reference implementation, `main` ALWAYS builds
a BVH from the objects in the world before rendering.